+++ src/main/java/soma/ghostrunner/domain/member/application/MemberService.java
@@ -3,10 +3,10 @@ package soma.ghostrunner.domain.member.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import soma.ghostrunner.domain.member.*;
 import soma.ghostrunner.domain.member.domain.Member;
 import soma.ghostrunner.domain.member.domain.MemberBioInfo;
-import soma.ghostrunner.domain.member.MemberNotFoundException;
+import soma.ghostrunner.domain.member.exception.InvalidMemberException;
+import soma.ghostrunner.domain.member.exception.MemberNotFoundException;
 import soma.ghostrunner.domain.member.dao.MemberRepository;
 import soma.ghostrunner.domain.member.api.dto.ProfileImageUploadRequest;
 import soma.ghostrunner.domain.member.application.dto.MemberCreationRequest;
