+++ src/main/java/soma/ghostrunner/domain/member/exception/MemberNotFoundException.java
@@ -1,4 +1,4 @@
-package soma.ghostrunner.domain.member;
+package soma.ghostrunner.domain.member.exception;
 
 import soma.ghostrunner.global.error.ErrorCode;
 import soma.ghostrunner.global.error.exception.EntityNotFoundException;
