+++ src/test/java/soma/ghostrunner/domain/auth/application/AuthServiceTest.java
@@ -12,7 +12,7 @@ import soma.ghostrunner.domain.auth.api.dto.response.AuthenticationResponse;
 import soma.ghostrunner.domain.auth.application.dto.JwtTokens;
 import soma.ghostrunner.domain.auth.exception.TokenTheftException;
 import soma.ghostrunner.domain.auth.resolver.impl.FirebaseUidResolver;
-import soma.ghostrunner.domain.member.InvalidMemberException;
+import soma.ghostrunner.domain.member.exception.InvalidMemberException;
 import soma.ghostrunner.domain.member.domain.Member;
 import soma.ghostrunner.domain.member.dao.MemberRepository;
 import soma.ghostrunner.domain.member.application.MemberService;
